+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.5.0 (2023-09-28)
+
+### Features
+
+- introducing ancestors queries to query for parents of an entity, the parents of parents, and so on:
+  - `Ecspanse.Query.select/2` new option: `:for_ancestors_of`
+  - `Ecspanse.Query.list_ancestors/1`
+  - `Ecspanse.Query.list_tagged_components_for_ancestors/2`
+
 ## v0.4.0 (2023-09-17)
 
 ### Breaking